Singly linked list.
My previous implementation of the linked list was poorly structured and
hard to work with, I will rebuild it from scratch with type hints in
mind, decoupling and correctness in mind. For now, I have implemented
the first of four types of the linked list, the forward linked list.

Improved tests will also be written, better than the existing ones which
I have deleted.

Types:
1. Singly linked list
2. Doubly linked list
3. Circular singly linked list
4. Circular doubly linked list