Implemented a clean linked list design.
Last time I tried implementing the linked list proved to be very hard
mainly dues to poor design, today's design is somewhat better. Used
abstract class for the all my linked list, then used it to add operators
to it, the adapter design pattern. Inherit from this adapter and
in=mplement the linked list logic and you end up with a working linked
list that is pretty much the same as the default python list.
It can be passed as an instance of the typing.MutableSequence in which
list is also a type of.
For now, I have implemented the Singly and Doubly linked list, the
Circular version will be added by the end of the week, hopefully, and
type hints for some other day.