Implemented a clean Linkedlist, finally.
Design is hard to get right, it is even harder when you know nothing. I
have reimplemented the linkedlist a dozen times and this is the first
time it seems to have gone well and reusable with a minimal interface,
hope this is reusable. And now, to the next data structure.