Implemented a basic set.
As a proof of concept, I have implemented a basic set that supports
adding, removing and checking for values on the set. Later, I will add
one that works and implements the typing.MutableSet along with the
HashTable.
The current implementation passes all tests and I believe this is a
green light. I'm trying to avoid reimplementations, they consume more
time and resources than the good they bring. Look at the linked list for
example.