Tried implementing ty.MutableSet
Unlike ty.MytableSequence, ty.MutableSet is going to be harder to
implement due to set complexity, ie, set operations are kinda getting
weird to implement, what is symetric_difference in python set object,
and why is it not part of ty.MutableSet. I donno what is what, but I'm
gonna need sometime to look at this python set thing.