Added some hash table tests and fixed some bugs
There was a subtle syntax error bug that passed when I added the __str__
method in the HashTable object, Itwas an invalid format string which I
did not even intend to put there, the unexpected colon was  to blame,
this has been fixed and in addition, the method `fromkeys` has also been
made available. Though I have struggled to get the method's type hint
correct, I will fix this later. For the tests, I am impressed that the
expansion operation works even for my hash table, I have a theory, any
object that yields a two item iterable can be used, that is, the first
index 0, should hold a string, while the second being the value to be
passed. I do not know if there is an implicit conversion of the key to a
string or this is an error. I will test this out and post the results
back here as a commit.