Read on binary trees and traversal
Simply a binary tree is a tree with atmost 2 children per node, thats
it. Unfortunately, there are lots of different types of binary trees out
there from full, completed to left/right-skewed and other ones like
binary search tree and AVL trees and the ad thing about all of them is
they are fucking binary trees. Except for the trees with special
properties like BST and AVL, a type of Self Balancing BST I think, the
rest I will definately not implement as it would be a great waste of
time and space doing repetitive work, maybe a tiny impl but not more
than that, the real joy and fulfillment will come from the self
balancing and B-* trees. I wonder how databases use these B-trees and
B+trees and red-black trees to store data, they are the hot cake.
Otherwise, I will document the traversals and types of Binary trees
later, it was fun playing around with the traversals. NIGHT.