That's all on Bloom Filters
On second thought, the tracking a bloom filters is just hard. If I count
on every insert then I may count duplicates, if I only count values
which are not in the bloom during insertion then I may not count values
that are not in the bloom but whose identifying bits are on hence will
not be counted especially as the capacity grows the counting becomes
more errorneous. Did some research on this kind of problem and got a
somewhat satisfying answer, hyperloglog, I will read and try to
implement this later to fix my tiny counting problem. And also, solving
the counting problem will help fix the analysis; confidence.

For now, I've stripped the Bloom Filter down to the simple good ol'
basic Bloom filter, it works perfectly. Night.