Cleaned up the BitPool and added tests
My designs are not state of the art but I'm willing to improve on them,
as practice, I thought redesigning the BitPool API and making it more
abstract using the protocals and mixins. From my understanding, a
protocal is a set of rules that should be adhered to in some defined
situation while mixins are broken implementations of some protocal or
system which cannot function independently. Mixins are almost like the
`Template design pattern`, they define or introduce definitions which
they use or are supposed to be implemented by the subclass or will be
found there. Mixins allow for some good abstarction and code reuse, they
can help eliminate inheritance or at least keep it in check and most
importantly, they carry implementation details pretty well.

By thinking in terms of properties an object has and what it needs
during system design, you get the benefit of a loose flexible system.
Consider the design of python and C++. Unlike C++ which forces you to
only have certain statements in a class body, python allows any type of
statement in a class body. The class body executes in some namespace and
executed declarations add new names to that namespace while the ones
that are not reached don't add. Although C++ is meant to be compiled
while python is interpreted, I still like the idea of keeping it simple
and not adding special rules for edge cases or atleast what we think
should be considered simple.

This type of work belongs to the `code` repo but `bitpool` seemed like a
good target for code design and refactoring. Intruth, I've learned a lot
of things from this especially on python's operator overloads like what
is the purpose for `__index__` and applying overloads and also
implementing them. Night.