Mixins flexibility
There was a bug in the index validator which made it spit out invalid
indices, the bug has been patched in a very non-intuitive way but it
works correctly. The best part about the fix is that by fixing the mixin
bug, I kinda fixed all objects that depend on it, mind-blown. The
distributive-ness of mixins are where its power lies. Actually, the
proper term is "Traits design pattern." Traits are a type of mixin, in
rust, a trait only holds functionality and lists out its dependencies
such that the implementor object meets the dependencies then it can use
the functions defined in the Trait; the contract. Mixins are not very
far from traits and to be general, they are traits to.