Implemented Multidimensional Heap
To make things simpler, a min heap can be used in place of max heap by
inversion, and vice versa. Therefore, implementing a min-heap has to be
the way to go. My earlier implementation of the heap was very cumbersome
and full of comparators and all these customization options which are
not really necessary, this time, no such mistake will be done.
Otherwise, there is a functioning multidimmensional heap which was a fun
read.